Ensure that ongoing services with an unknown EndDate are available in the immediate future
Services may run for long time unchanged. A current example of this is the Celtic Coaches X47 (Wales), with an `OperatingPeriod StartDate` of 2020-06-29 and no defined `EndDate`.

The proposed change should ensure that future services are available for a year after they start, and that current services are available for a year from the time of import.